+++ server/src/resolvers/objectTypes.ts
@@ -16,4 +16,7 @@ export class UserResponse {
 
   @Field(() => User, { nullable: true })
   user?: User;
+
+  @Field(() => String, { nullable: true })
+  token?: string;
 }